Adapted routines for estimating contamination for tileDeck files
This is also working under MPI too. However, the noise part of the
simulations needs more work, in order to reproduce RMS maps that
look like the real data.

The sky sim part is now using enlib (orders of magnitude faster
than flipper equivalent), so that is now a dependency.