Relativistic corrections for multi-freq filter
We now write a datacube of relative weights by frequency, so we can calculate
the relativistic correction as the weighted average of the relative corrections
for each frequency, according to how much they contribute to y0~. These weights
are also written into output catalogs as fixed_y_c_weight_obsFreqGHz. This
provides a way of quicky checking which parts of the catalog come from single
frequency only observations.

Modified parsing of the config file, such that it's no longer necessary to give
the saveRMSMap and saveFreqWeightsMap keys in the filter settings for the entry
corresponding to photFilterLabel: since these are always needed, we may as well
add them if the user forgets to.

Updated E-D56 and AdvACT examples accordingly