Fixed bug in MockSurvey.update and added checking of map, point source mask sizes
MockSurvey.update now updates the interpolators used for drawing samples
when the cosmological parameters are changed.

In map pre-processing stage, we now check whether the point source mask
and survey mask are the same dimensions as the input map - if they are
not, we throw an exception and print a useful error message.