Add mask validity check and MPI Abort on exception
Will now catch masks with negative values on start up (such masks can
give e.g., negative area).

Replaced sys.excepthook to trigger MPI abort when exception encountered.
This is because mpi4py will never do so (which meant that when a
process died, nemo never completed and wasted cycles). The drawback
of doing this is that the full traceback isn't printed, but at least
we get the error message itself and kill all the processes.