No longer filling point source holes with white noise
Just smoothing, no longer adding noise. This should have been spotted
earlier... but adding random noise to something that ends up in the
filter is bad for repeatability (even if qualitatively it doesn't have
much impact). The main impact is on regions next to point source holes.
Which is our primary source of artifacts (holes not being big enough).