Removed DR5 hard coded deltaT PS cuts
These were not the issue with the DR5 public pre-release: that
turned out to be a bug in the old version related to filling mask
holes (median filter applied N times for N catalogs in the
config file). This only affected SNR of a few clusters near point
source holes (SNR was overestimated in old version).