New tests and change of output dir path
Working on simpler, easier to set-up tests. Starting with a set of
'quick' tests that run the quickstart tutorial.

To get this working, changed how the output directory is set from
Nemo, when given just a config file: this is now created in the
current working directory, allowing config files from anywhere to
be used with relative paths. Paths given in config files are taken
to be relative to the current working directory, not relative to
where the config file lives.