Fix pathological 1/f noise issue
By default, 1/f noise becomes pathologically high as you approach ell=0.  This creates unrealistic maps.  In practice, the real atmosphere and the map-maker transfer function lead to zeroing of the noise on large scales. I arbitrarily set the noise for  lmin<100 (default) to be zero to avoid this, and provide lmin as a parameter to simMap.